refactor(test): localize #507 orphan-widget disposal to test_scan_dialog
Move the deferred-delete teardown fixture out of the global conftest into test_scan_dialog.py (module-scoped autouse) so it disposes only the bare _SourceListWidget/ScanDialog orphans this file leaks. A future widget leak in another module then still surfaces instead of being globally swept; conftest.py returns to its master state. Verified 5x/5x green full-suite on 3.12-local.
